[Decode] Fix Coverity static analysis issues in decode component
Fix 11 Coverity defects in the Decode component across MJPEG, JXS,
AV1, H264, and H265 decoders.
Also fix scan-modified-class: use RT_FUNC_ENTER(CheckGUID) and add
OFFSET_FUNC_CheckGUID to mfx_smartv_trace_offset.h.

MISSING_LOCK (data race condition) fixes:
- CID 3583175: Add m_mGuard lock to VideoDECODEMJPEG::Reset(), which
  also covers IsSameVideoParam() accessing m_response without lock
- CID 3583205: Same Reset() lock fix covers m_platform access
- CID 3583326: Add m_mGuard lock to JXS decoder Reset() for
  m_core access
- CID 3583328: Add m_mGuard lock to JXS decoder DecodeFrameCheck()
  for m_core access
- CID 3583455: Add guard lock to AV1Decoder::GetCurrFrame() for
  lastest_submitted_frame access
- CID 3583545: Same JXS Reset() lock fix covers m_platform access

INTEGER_OVERFLOW fixes:
- CID 3515335: Guard against l==0 in read_uniform() to prevent
  uint32_t underflow on l-1 when GetUnsignedBits(n)==0
- CID 3535741: Change loop variable sliceId from uint32_t to int32_t
  in H265DecoderFrameInfo::EliminateErrors() to avoid wrap-around
  overflow when resetting to -1 before continue
- CID 3541793: Change loop variable sliceId from size_t to ptrdiff_t
  in SetOfSlices::CleanUseless() to avoid underflow on --sliceId
- CID 3553573: Change loop variable pos from size_t to ptrdiff_t in
  AccessUnit::CleanUseless() to avoid underflow on pos--
- CID 3556511: Add explicit (int32_t) cast in
  DecodeExpGolombOne_H264_1u32s() when assigning uint32_t sentinel
  value 0xFFFFFFFF to int32_t *pDst

+++ _studio/shared/umc/codec/h264_dec/include/umc_h264_bitstream_headers.h
@@ -391,7 +391,7 @@ inline bool DecodeExpGolombOne_H264_1u32s (uint32_t **ppBitStream,
         if ((bool)(isSigned) == true)
             *pDst = (sval & 1UL) ? (int32_t)((sval >> 1) + 1) : -(int32_t)(sval >> 1);
         else
-            *pDst = sval;
+            *pDst = (int32_t)sval;
     }
    return true;
 }
